Update NOTICE.md and architecture.md; enhance archive engine documentation and add new dependencies. Modify README.md to clarify built-in viewer capabilities and automation features. Regenerate Riverpod hashes in create_container_controller.g.dart and usb_create_container_controller.g.dart. Update unit test for UsbMassStorageDevice to ensure correct CBW size.

+++ android/app/src/test/kotlin/com/aeidolon/vaultexplorer/usb/UsbMassStorageDeviceBuildCbwTest.kt
@@ -29,7 +29,7 @@ class UsbMassStorageDeviceBuildCbwTest {
     @Test
     fun `total length is always the fixed 31-byte CBW size`() {
         assertEquals(31, UsbMassStorageDevice.buildCbw(1, byteArrayOf(0x25), 0, dirIn = true).size) // READ CAPACITY(10), 1-byte CDB
-        assertEquals(31, UsbMassStorageDevice.buildCbw(1, ByteArray(16) { 0x88 }, 512, dirIn = true).size) // full 16-byte CDB, e.g. READ(16)
+        assertEquals(31, UsbMassStorageDevice.buildCbw(1, ByteArray(16) { 0x88.toByte() }, 512, dirIn = true).size) // full 16-byte CDB, e.g. READ(16)
     }
 
     @Test
